OrderSphere/requirements.txt
Flask==3.0.3
mysql-connector-python==8.3.0
Werkzeug==3.0.3
python-dotenv==1.0.0
